[Feature] 유저별 신청 리스트 조회 API추가, 각 API별 권한 인증 기능 추가

+++ src/domains/challenges/challenges.controller.ts
@@ -125,7 +125,7 @@ import {
  *                 rejectedReason: null
  *                 approvalStatus: "PENDING"
  *       404:
- *         description: 챌린지를 찾을 수 없음
+ *         description: 요청한 리소스를 찾을 수 없습니다.
  *       500:
  *         description: 서버 오류
  */
@@ -219,7 +219,7 @@ const getChallenge: GetController<
  *                   type: integer
  *                   description: 총 챌린지 수
  *       403:
- *         description: 관리자 권한 없음
+ *         description: 권한 없음
  *       500:
  *         description: 서버 오류
  */
@@ -251,53 +251,45 @@ const getChallengeListByAdmin: GetController<
 
 /**
  * @swagger
- * /api/challenges:
+ * /api/challenges/user:
  *   get:
- *     summary: Retrieve a list of challenges
- *     description: Fetch a paginated list of challenges based on filters such as document type, fields, approval status, and keyword.
  *     tags:
  *       - Challenges
+ *     summary: 사용자별 챌린지 목록 조회
+ *     description: 사용자가 생성한 챌린지 목록을 조회합니다.
  *     parameters:
  *       - in: query
- *         name: documentType
+ *         name: orderBy
  *         schema:
  *           type: string
- *           enum: [TYPE_A, TYPE_B, TYPE_C] # Replace with actual DocumentType enum values
- *         description: Filter challenges by document type
- *       - in: query
- *         name: fields
- *         schema:
- *           type: array
- *           items:
- *             type: string
- *             enum: [FIELD_X, FIELD_Y, FIELD_Z] # Replace with actual FieldType enum values
- *         description: Filter challenges by fields
+ *           enum: [applyFirst, applyLast, deadLineFirst, deadLineLast]
+ *         description: 챌린지 정렬 기준
  *       - in: query
  *         name: approvalStatus
  *         schema:
  *           type: string
- *           enum: [PENDING, APPROVED, REJECTED] # Replace with actual approval status values
- *         description: Filter challenges by approval status
+ *           enum: [PENDING, APPROVED, REJECTED]
+ *         description: 승인 상태로 필터링
  *       - in: query
  *         name: keyword
  *         schema:
  *           type: string
- *         description: Search challenges by title or description
+ *         description: 제목 또는 설명에서 키워드 검색
  *       - in: query
  *         name: page
  *         schema:
  *           type: integer
  *           default: 1
- *         description: Page number for pagination
+ *         description: 페이지 번호
  *       - in: query
  *         name: limit
  *         schema:
  *           type: integer
  *           default: 10
- *         description: Number of challenges per page
+ *         description: 페이지당 항목 수
  *     responses:
  *       200:
- *         description: A list of challenges and the total count
+ *         description: 성공적으로 사용자별 챌린지 목록 반환
  *         content:
  *           application/json:
  *             schema:
@@ -308,25 +300,159 @@ const getChallengeListByAdmin: GetController<
  *                   items:
  *                     type: object
  *                     properties:
+ *                       id:
+ *                         type: string
+ *                         description: 챌린지 ID
  *                       title:
  *                         type: string
- *                       field:
+ *                         description: 챌린지 제목
+ *                       approvalStatus:
  *                         type: string
- *                       maxParticipants:
- *                         type: integer
- *                       currentParticipants:
- *                         type: integer
- *                       deadline:
+ *                         description: 승인 상태
+ *                       createdAt:
  *                         type: string
  *                         format: date-time
- *                       documentType:
+ *                         description: 생성 날짜
+ *                       updatedAt:
  *                         type: string
+ *                         format: date-time
+ *                         description: 마지막 업데이트 날짜
  *                 totalCount:
  *                   type: integer
- *       400:
- *         description: Invalid request parameters
+ *                   description: 총 챌린지 수
+ *             example:
+ *               challenges:
+ *                 - id: "123e4567-e89b-12d3-a456-426614174000"
+ *                   title: "프론트엔드 번역 챌린지"
+ *                   approvalStatus: "APPROVED"
+ *                   createdAt: "2025-03-29T12:00:00.000Z"
+ *                   updatedAt: "2025-03-30T12:00:00.000Z"
+ *                 - id: "789e4567-e89b-12d3-a456-426614174001"
+ *                   title: "백엔드 번역 챌린지"
+ *                   approvalStatus: "PENDING"
+ *                   createdAt: "2025-03-28T12:00:00.000Z"
+ *                   updatedAt: "2025-03-29T12:00:00.000Z"
+ *               totalCount: 2
+ *       403:
+ *         description: 사용자 인증 실패
  *       500:
- *         description: Internal server error
+ *         description: 서버 오류
+ */
+const getChallengeListByUser: GetController<
+  never,
+  ChallengeRequestQueries,
+  GetChallengeListByAdminResponse
+> = async (req, res, next) => {
+  try {
+    const userId = req.user?.id;
+    const { orderBy, page = '1', limit = '10', approvalStatus, keyword } = req.query;
+    const result = await ChallengesService.getChallengeListByUser({
+      orderBy,
+      page,
+      limit,
+      approvalStatus,
+      keyword,
+      userId: userId as string,
+    });
+    res.status(200).send(result);
+  } catch (err) {
+    next(err);
+  }
+};
+
+/**
+ * @swagger
+ * /api/challenges:
+ *   get:
+ *     tags:
+ *       - Challenges
+ *     summary: 챌린지 목록 조회
+ *     description: 필터 및 정렬 조건을 기반으로 챌린지 목록을 조회합니다.
+ *     parameters:
+ *       - in: query
+ *         name: documentType
+ *         schema:
+ *           type: string
+ *           enum: [BLOG, OFFICIAL]
+ *         description: 문서 타입으로 필터링
+ *       - in: query
+ *         name: fields
+ *         schema:
+ *           type: string
+ *           enum: [NEXTJS, MODERNJS, API, WEB, CAREER]
+ *         description: 챌린지 분야로 필터링
+ *       - in: query
+ *         name: keyword
+ *         schema:
+ *           type: string
+ *         description: 제목 또는 설명에서 키워드 검색
+ *       - in: query
+ *         name: page
+ *         schema:
+ *           type: integer
+ *           default: 1
+ *         description: 페이지 번호
+ *       - in: query
+ *         name: limit
+ *         schema:
+ *           type: integer
+ *           default: 10
+ *         description: 페이지당 항목 수
+ *     responses:
+ *       200:
+ *         description: 성공적으로 챌린지 목록 반환
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 challenges:
+ *                   type: array
+ *                   items:
+ *                     type: object
+ *                     properties:
+ *                       id:
+ *                         type: string
+ *                         description: 챌린지 ID
+ *                       title:
+ *                         type: string
+ *                         description: 챌린지 제목
+ *                       documentType:
+ *                         type: string
+ *                         description: 문서 타입
+ *                       field:
+ *                         type: string
+ *                         description: 챌린지 분야
+ *                       createdAt:
+ *                         type: string
+ *                         format: date-time
+ *                         description: 생성 날짜
+ *                       updatedAt:
+ *                         type: string
+ *                         format: date-time
+ *                         description: 마지막 업데이트 날짜
+ *                 totalCount:
+ *                   type: integer
+ *                   description: 총 챌린지 수
+ *             example:
+ *               challenges:
+ *                 - id: "123e4567-e89b-12d3-a456-426614174000"
+ *                   title: "프론트엔드 번역 챌린지"
+ *                   documentType: "BLOG"
+ *                   field: "NEXTJS"
+ *                   createdAt: "2025-03-29T12:00:00.000Z"
+ *                   updatedAt: "2025-03-30T12:00:00.000Z"
+ *                 - id: "789e4567-e89b-12d3-a456-426614174001"
+ *                   title: "백엔드 번역 챌린지"
+ *                   documentType: "OFFICIAL"
+ *                   field: "API"
+ *                   createdAt: "2025-03-28T12:00:00.000Z"
+ *                   updatedAt: "2025-03-29T12:00:00.000Z"
+ *               totalCount: 2
+ *       400:
+ *         description: 잘못된 요청 데이터
+ *       500:
+ *         description: 서버 오류
  */
 
 const getChallengeList: GetController<
@@ -404,7 +530,7 @@ const getChallengeList: GetController<
  *                 description: 원본 문서 URL
  *                 example: "https://example.com/original-doc"
  *     responses:
- *       201:
+ *       200:
  *         description: 챌린지가 성공적으로 생성되었습니다.
  *         content:
  *           application/json:
@@ -475,6 +601,7 @@ const postChallenge: PostController<
       deadline,
       originURL,
     } = req.body;
+    const userId = req.user?.id;
     const result = await ChallengesService.createChallenge({
       title,
       description,
@@ -483,6 +610,7 @@ const postChallenge: PostController<
       maxParticipants,
       deadline,
       originURL,
+      userId: userId as string,
     });
     res.status(200).send({ challenge: result, code: 200 });
   } catch (err) {
@@ -597,6 +725,10 @@ const postChallenge: PostController<
  *                       example: "2025-03-30T12:00:00.000Z"
  *       400:
  *         description: 잘못된 요청 데이터
+ *       401:
+ *         description: 로그인 정보 없음
+ *       403:
+ *         description: 권한 없음
  *       404:
  *         description: 챌린지를 찾을 수 없음
  *       500:
@@ -609,6 +741,17 @@ const patchChallenge: PatchController<
 > = async (req, res, next) => {
   try {
     const id = req.params.challengeId;
+    const existChallenge = await ChallengesService.getChallenge(id);
+    const isEqualUser = req.user?.id === existChallenge.challenge?.userId;
+    const authRole = req.user?.role;
+    if(!existChallenge.challenge){
+      next({ status: 404 });
+      return;
+    }
+    if (authRole !== "ADMIN" && !isEqualUser) {
+      next({ status: 403 });
+      return;
+    }
     const {
       title,
       description,
@@ -662,6 +805,10 @@ const patchChallenge: PatchController<
  *                   type: integer
  *                   description: 응답 코드
  *                   example: 200
+ *       401:
+ *         description: 로그인 정보 없음
+ *       403:
+ *         description: 권한 없음
  *       404:
  *         description: 삭제할 챌린지를 찾을 수 없음
  *         content:
@@ -700,6 +847,17 @@ const deleteChallenge: DeleteController<
 > = async (req, res, next) => {
   try {
     const id = req.params.challengeId;
+    const existChallenge = await ChallengesService.getChallenge(id);
+    const isEqualUser = req.user?.id === existChallenge.challenge?.userId;
+    const authRole = req.user?.role;
+    if(!existChallenge.challenge){
+      next({ status: 404 });
+      return;
+    }
+    if (authRole !== "ADMIN" && !isEqualUser) {
+      next({ status: 403 });
+      return;
+    }
     const result = await ChallengesService.deleteChallenge(id);
     if (!result) {
       next({ statusCode: 404 });
@@ -713,6 +871,7 @@ const deleteChallenge: DeleteController<
 
 const ChallengesController = {
   getChallengeList,
+  getChallengeListByUser,
   getChallengeListByAdmin,
   getChallenge,
   postChallenge,
